+++ R/brmsformula.R
@@ -1215,6 +1215,62 @@ update.mvbrmsformula <- function(object, formula., ...) {
   object
 }
 
+#' Update Formula Addition Terms
+#'
+#' Update additions terms used in formulas of \pkg{brms}. See
+#' \code{\link{addition-terms}} for details.
+#'
+#' @param formula Two-sided formula to be updated.
+#' @param adform One-sided formula containing addition terms to update
+#'   \code{formula} with.
+#' @param action Indicates what should happen to the existing addition terms in
+#'   \code{formula}. If \code{"update"} (the default), old addition terms that
+#'   have no corresponding term in \code{adform} will be kept. If
+#'   \code{"replace"}, all old addition terms will be removed.
+#'
+#' @return An object of class \code{formula}.
+#' 
+#' @examples 
+#' form <- y | trials(size) ~ x
+#' update_adterms(form, ~ trials(10))
+#' update_adterms(form, ~ weights(w))
+#' update_adterms(form, ~ weights(w), action = "replace")
+#' update_adterms(y ~ x, ~ trials(10))
+#'
+#' @export
+update_adterms <- function(formula, adform, action = c("update", "replace")) {
+  formula <- as.formula(formula)
+  adform <- as.formula(adform)
+  action <- match.arg(action)
+  if (is.null(lhs(formula))) {
+    stop2("Can't update a ond-sided formula.")
+  }
+  str_formula <- formula2str(formula)
+  old_ad <- get_matches("(?<=\\|)[^~]*(?=~)", str_formula, perl = TRUE)
+  new_ad_terms <- attr(terms(adform), "term.labels")
+  if (action == "update" && length(old_ad)) {
+    # extract adterms from the original formula
+    old_ad <- formula(paste("~", old_ad))
+    old_ad_terms <- attr(terms(old_ad), "term.labels")
+    old_adnames <- get_matches("^[^\\(]+", old_ad_terms)
+    old_adnames <- sub("^resp_", "", old_adnames)
+    new_adnames <- get_matches("^[^\\(]+", new_ad_terms)
+    new_adnames <- sub("^resp_", "", new_adnames)
+    # keep unmatched adterms of the original formula
+    keep <- !old_adnames %in% new_adnames
+    new_ad_terms <- c(old_ad_terms[keep], new_ad_terms)
+  }
+  if (length(new_ad_terms)) {
+    new_ad_terms <- paste(new_ad_terms, collapse = "+")
+    new_ad_terms <- paste("|", new_ad_terms)
+  }
+  resp <- gsub("\\|.+", "", deparse_combine(formula[[2]]))
+  out <- formula(paste(resp, new_ad_terms, "~1"))
+  out[[3]] <- formula[[3]]
+  attributes(out) <- attributes(formula)
+  out
+}
+
 #' @export
 print.brmsformula <- function(x, wsp = 0, digits = 2, ...) {
   cat(formula2str(x$formula, space = "trim"), "\n")